requirements.txt
fastapi
minio
uvicorn 
python-dotenv
python-multipart